style: move pkg/middleware to pkg/core/middleware
