+++ pkg/util/ctxutil/ctxutil.go
@@ -17,7 +17,7 @@ package ctxutil
 import (
 	"context"
 
-	"github.com/KusionStack/karbour/pkg/middleware"
+	"github.com/KusionStack/karbour/pkg/core/middleware"
 	"github.com/go-logr/logr"
 	"k8s.io/klog/v2"
 )
